+++ modules/host/mdm/instructionreadsector.go
@@ -122,10 +122,22 @@ func (i *instructionReadSector) Cost() (types.Currency, types.Currency, error) {
 	if err != nil {
 		return types.ZeroCurrency, types.ZeroCurrency, err
 	}
-	return ReadCost(i.staticState.priceTable, length), types.ZeroCurrency, nil
+	cost, refund := ReadCost(i.staticState.priceTable, length)
+	return cost, refund, nil
+}
+
+// Memory returns the memory allocated by the 'ReadSector' instruction beyond
+// the lifetime of the instruction.
+func (i *instructionReadSector) Memory() uint64 {
+	return ReadMemory()
 }
 
 // ReadOnly for the 'ReadSector' instruction is 'true'.
 func (i *instructionReadSector) ReadOnly() bool {
 	return true
+}
+
+// Time returns the execution time of a 'ReadSector' instruction.
+func (i *instructionReadSector) Time() uint64 {
+	return TimeReadSector
 }