+++ CHANGELOG.md
@@ -13,8 +13,11 @@ Latest:
 **Key Updates**
  - Add `--root` parameter to `siac renter delete` that allows passing absolute
    instead of relative file paths.
+ - Add `siac skynet unpin` subcommand.
 
 **Bugs Fixed**
+ - Fixed `siac skynet ls` not working when files were passed as input. It is now
+   able to access specific files in the Skynet folder.
 
 **Other**